fixes for IBM i - mostly clarifying ambiguous casts.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/10379


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@34274 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/plurrule.cpp
@@ -1475,7 +1475,7 @@ int64_t FixedDecimal::getFractionalDigits(double n, int32_t v) {
       case 2: return (int64_t)(fract*100.0 + 0.5);
       case 3: return (int64_t)(fract*1000.0 + 0.5);
       default:
-          double scaled = floor(fract * pow(10.0, v) + 0.5);
+          double scaled = floor(fract * pow(10.0, (double)v) + 0.5);
           if (scaled > U_INT64_MAX) {
               return U_INT64_MAX;
           } else {
